requirements.txt
openai==0.27.2
streamlit==1.32.2
langchain==0.0.251
python-dotenv==1.0.1